Implements getting user messages from the back-end
In order to increase the bedth of data that can be retrieved from the
database, this commit implements the ability to retrieve user messages
from the back-end.

+++ backend/src/response_models/paginatied_response.rs
@@ -1,15 +1,15 @@
 #[derive(Debug, serde::Serialize)]
 pub struct PaginatedResponse<T> {
   pub item: T,
-  pub total_items: usize,
-  pub total_pages: usize,
+  pub total_items: u64,
+  pub total_pages: u64,
 
   #[serde(default = "PaginatedResponse::default_page_size")]
-  pub page_size: usize,
+  pub page_size: u64,
 }
 
 pub trait DefaultPagination {
-  fn default_page_size() -> usize {
+  fn default_page_size() -> u64 {
     100
   }
 }
